+++ src/components/rootPage/index.js
@@ -1,6 +1,7 @@
 export { default as Banner } from './Banner';
 export { default as Board } from './Board';
 export { default as Posters } from './Posters';
+export { default as Poster } from './Poster';
 
 export { default as Category } from './Category';
 export { default as Slide } from './Slide';
